requirements.txt
requests
Jinja2
python-dotenv
